Option to delete already existing website
Program checks is selected path to save the website already exists. If it does it asks the user to delete it and proceed. If the user does not want to delete it, the program will prompt for a new path and will repeat until that path is okay to write to.